// src/generated/api/models/PostDto.ts
export interface PostDto {
    /**
     * 
     * @type {number}
     * @memberof PostDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof PostDto
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof PostDto
     */
    content?: string;
    /**
     * 
     * @type {string}
     * @memberof PostDto
     */
    type?: PostDtoTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof PostDto
     */
    imageUrl?: string;
    /**
     * 
     * @type {boolean}
     * @memberof PostDto
     */
    active?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof PostDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof PostDto
     */
    updatedAt?: Date;
    /**
     * 
     * @type {UserSummaryDto}
     * @memberof PostDto
     */
    author?: UserSummaryDto;
    /**
     * 
     * @type {number}
     * @memberof PostDto
     */
    likesCount?: number;
    /**
     * 
     * @type {number}
     * @memberof PostDto
     */
    commentsCount?: number;
    /**
     * 
     * @type {boolean}
     * @memberof PostDto
     */
    likedByCurrentUser?: boolean;
}


/**
 * @export
 */
export const PostDtoTypeEnum = {
    Public: 'PUBLIC',
    Private: 'PRIVATE',
    Draft: 'DRAFT'
} as const;
export type PostDtoTypeEnum = typeof PostDtoTypeEnum[keyof typeof PostDtoTypeEnum];